update czctrl state watch to file notify

+++ crates/czctrl/src/config.rs
@@ -1,7 +1,7 @@
 // apply
 pub const DEFAUL_LIBVIRT_URI: &str = "qemu:///system";
 pub const TRY_COUNT: i8 = 10;
-pub const TRY_INTERVAL: u64 = 1; // try interval (second)
+pub const TRY_INTERVAL: i8 = 1; // try interval (second)
 
 // observe
 pub const RESCTL_ROOT: &str = "/sys/fs/resctrl";
